Issue SB-12311 BaseContentManager added, refactored contentMgr
