--- content-service/content-api-core/src/main/scala/org/ekstep/managers/CreateContentMgr.scala
@@ -1,7 +1,0 @@
-package org.ekstep.managers
-
-import org.ekstep.common.mgr.BaseManager
-
-class CreateContentMgr extends BaseManager{
-
-}
